+++ lib/src/header/header.dart
@@ -84,7 +84,9 @@ abstract class RefreshHeaderState<T extends RefreshHeader> extends State<T> {
 /// Header监听器
 abstract class HeaderListener {
   // 更新视图高度
+  @deprecated
   void updateHeight(double newHeight){}
+  void updateHeaderHeight(double newHeight){}
   // 回调开始刷新方法
   void onRefreshStart(){}
   // 回调准备刷新方法
@@ -164,6 +166,7 @@ class _ListenerHeaderState extends RefreshHeaderState<ListenerHeader> {
   void updateHeight(double newHeight) {
     super.updateHeight(newHeight);
     widget.listener.updateHeight(newHeight);
+    widget.listener.updateHeaderHeight(newHeight);
   }
 }