+++ lib/bezier_bounce_footer.dart
@@ -75,8 +75,8 @@ class BezierBounceFooterState extends RefreshFooterState<BezierBounceFooter> wit
 
   // 加载结束
   @override
-  Future onLoadEnd() async {
-    super.onLoadEnd();
+  Future onLoadClose() async {
+    super.onLoadClose();
     setState(() {
       showBounce = false;
     });